refactor(migration): isolate reusable release lifecycle
